+++ src/app/app.component.scss
@@ -532,59 +532,6 @@
     }
 }
 
-.rfid-panel {
-    display: block;
-    position: fixed;
-    bottom: 2rem;
-    right: 2rem;
-    // min-width: 10rem;
-    // min-height: 20rem;
-    // width: 50rem;
-    max-width: 90%;
-    background: linear-gradient(45deg, $color-grey-4, $color-grey-1);
-    border-radius: 3rem;
-    border: .5rem solid $color-primary;
-    z-index: 30;
-    padding: 2rem;
-    box-shadow: 0 0 5rem $color-grey-2, inset 0 0 2rem $color-grey-2;
-    &>div {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-    }
-    i {
-        // height: 5rem;
-        // width: 5rem;
-        padding: 1rem;
-        color: white;
-        &:hover {
-            color: $color-primary;
-            cursor: pointer;
-        }
-    }
-    .buttons-row {
-        i {
-            margin: 1rem;
-        }
-        font-size: 5rem;
-    }
-    p {
-        color: rgba($color-grey-0, 0.8);
-        font-size: 2rem;
-    }
-    a {
-        &:hover {
-            &,
-            &>* {
-                color: $color-primary;
-                cursor: pointer;
-            }
-        }
-        font-size: 2.3rem;
-    }
-}
-
 @keyframes pulse-yellow {
     0% {
         border: 5px solid $color-grey-1;
