+++ src/app/competitor-zone/referee/add-fight-result/add-fight-result.component.scss
@@ -8,7 +8,6 @@
     align-items: center;
     flex-direction: column;
     width: 100%;
-    text-align: center;
     app-input {
         width: 100%;
         max-width: 50rem;
@@ -88,11 +87,4 @@ h5 {
 
 .result {
     max-width: min-content !important;
-}
-
-@media only screen and (max-width: $max-width-4) {
-    .grid {
-        display: flex;
-        flex-direction: column;
-    }
 }
