+++ src/app/competitor-zone/referee/add-time-result/add-time-result.component.html
