--- src/app/competitor-zone/referee/announcements/announcements.component.html
@@ -1,12 +1,0 @@
-<div class="router-child">
-    <section class="announcements">
-        <div [formGroup]="formMessage" *ngIf="userSerceice.isAdmin" class="announcements-form">
-            <textarea formControlName="message" placeholder="Wiadomość do wszystkich użytkowników"></textarea>
-            <button [disabled]="!isFormMessageValid" class="submit-button" (click)="sendMessage()">{{'competitor-zone.competitor.send' | translate}}</button>
-        </div>
-    </section>
-    <section *ngFor="let message of getAnnouncements" class="announcements-one">
-        <label>{{message.tresc}}</label>
-        <span><b>{{message.czas_nadania | date: 'HH:mm:ss'}}</b>{{message.czas_nadania | date: 'dd/MM/YYYY'}}</span>
-    </section>
-</div>
