--- src/app/competitor-zone/referee/announcements/announcements.component.scss
@@ -1,79 +1,0 @@
-@import '../../../scss/colors';
-@import '../../../scss/variables';
-.announcements {
-    width: 100%;
-    max-width: 100%;
-    display: grid;
-    &-one {
-        display: flex;
-        flex-direction: row;
-        width: 100%;
-        padding: 2rem;
-        border-bottom: .1rem dashed rgba($color-grey-0, 0.5);
-        label {
-            font-family: 'Jura';
-            font-size: 2rem;
-            width: 100%;
-        }
-        span {
-            font-family: 'Goldman';
-            b {
-                font-size: 1.3rem;
-            }
-            font-size: 1rem;
-            display: flex;
-            flex-direction: column;
-            text-align: center;
-            width: max-content;
-        }
-    }
-    &-form {
-        padding: 2rem !important;
-        width: 100%;
-        max-width: none;
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-        p,
-        label,
-        &>div {
-            width: 100%;
-            max-width: 100%;
-        }
-        &:nth-child(2) {
-            grid-column: 2;
-        }
-        &:last-child {
-            padding: 0 2rem;
-            width: 100%;
-            height: 100%;
-            justify-content: center;
-            align-items: center;
-            textarea {
-                min-height: 20rem;
-                width: 100%;
-                height: 100%;
-                resize: none;
-                background-color: transparent;
-                border: 2px solid $color-grey-0;
-                border-radius: 2rem;
-                padding: 1rem;
-                font-family: $title-font;
-                font-weight: bold;
-                font-size: 2rem;
-                color: $color-white;
-            }
-            textarea::placeholder {
-                color: $color-grey-0;
-            }
-            button {
-                font-size: 1.5rem;
-                margin: 0;
-                margin-top: 2rem;
-                padding: 2rem !important;
-                width: 100%;
-                height: min-content;
-            }
-        }
-    }
-}
