+++ src/app/competitor-zone/referee/competitors/competitor/competitor.component.scss
@@ -1,7 +1,6 @@
 @import '../../../user/my-robots/robot/robot.component.scss';
 @import '../../../user/settings/settings.component.scss';
 @import '../../../../scss/variables';
-@import '../../positions/positions.component.scss';
 .competitor-name {
     width: max-content;
     padding: 1rem 0!important;
@@ -253,9 +252,4 @@
 
 .banned {
     color: red !important;
-}
-
-.grid-button {
-    margin: 0 !important;
-    margin-top: 1rem !important;
 }
