+++ src/app/competitor-zone/referee/competitors/competitor/competitor.component.ts
@@ -9,8 +9,6 @@ import { UserService } from 'src/app/services/user.service';
 import { UiService } from 'src/app/services/ui.service';
 import { TranslateService } from '@ngx-translate/core';
 import { Title } from '@angular/platform-browser';
-import { PositionsService } from 'src/app/services/positions.service';
-import { Esp32Service } from '../../../../services/esp32.service';
 
 
 @Component({
@@ -26,22 +24,15 @@ export class CompetitorComponent {
   public formPostal: FormGroup;
   public formMessage: FormGroup;
   public formUserType: FormGroup;
-  public formReferee: FormGroup;
-
   private subs: Subscription = new Subscription;
   public user: any = null;
   public user_roboty: any = null;
   private loading = false;
   private loadingStarerpack = false;
-  public loadingReferee: boolean = false;
-
-  public allPositions: Array<any> | null = null;
-  public allReferePositions: Array<any> = [];
-
 
   constructor(private route: ActivatedRoute, private formBuilder: FormBuilder, public authService: AuthService,
-    private categoriesService: CategoriesService, public userSerceice: UserService, private router: Router, public esp32Service: Esp32Service,
-    private ui: UiService, private translate: TranslateService, private refereeService: RefereeService, private titleService: Title, private positionsService: PositionsService) {
+    private categoriesService: CategoriesService, public userSerceice: UserService, private router: Router,
+    private ui: UiService, private translate: TranslateService, private refereeService: RefereeService, private titleService: Title) {
     const uzytkownik_uuid = this.route.snapshot.paramMap.get('uzytkownik_uuid');
     this.authService.getRegisterAddons();
 
@@ -54,13 +45,9 @@ export class CompetitorComponent {
     this.formUserType = this.formBuilder.group({
       user_type: [null, [Validators.required]]
     });
-    this.formReferee = this.formBuilder.group({
-      position: [null, [Validators.required]]
-    });
+    const sub1 = combineLatest(this.categoriesService.categories$, this.refereeService.allUsers$).subscribe(async (val) => {
 
-    const sub1 = combineLatest(this.categoriesService.categories$, this.refereeService.allUsers$, this.positionsService.allPositions$).subscribe(async (val) => {
-
-      if (val[0] !== null && val[1] !== null && val[2] !== null) {
+      if (val[0] !== null && val[1] !== null) {
         const categories = val[0];
         const users: Array<any> = JSON.parse(JSON.stringify(val[1]));
         const user = users.find(user => user.uzytkownik_uuid === uzytkownik_uuid)
@@ -71,18 +58,6 @@ export class CompetitorComponent {
             this.formUserType.controls['user_type'].setValue(this.user.uzytkownik_typ);
           }, 300)
           this.titleService.setTitle(`🧑 ${this.user.imie} ${this.user.nazwisko}`);
-          if (this.user.uzytkownik_typ >= 2) {
-            this.allPositions = val[2];
-            if (this.allPositions) {
-              this.allReferePositions = [];
-              this.allPositions.forEach(position => {
-                let sedziowie = this.stringToArray(position.sedziowie)
-                if (sedziowie.includes(this.user.uzytkownik_uuid)) {
-                  this.allReferePositions.push(position)
-                }
-              })
-            }
-          }
         }
       }
     });
@@ -123,31 +98,6 @@ export class CompetitorComponent {
     }
   }
 
-  async onRemoveReferee(stanowisko: any) {
-    if(this.user) {
-      const decision = await this.ui.wantToContinue(`Czy na pewno chcesz usunąć sędziego z stanowiska ${stanowisko.nazwa_stanowiska}?`)
-      if(!decision) return;
-      this.loadingReferee = true;
-      this.positionsService.removeRefereeFromPosition(stanowisko.stanowisko_id,this.user.uzytkownik_uuid).then(succes => {
-        this.ui.showFeedback("succes", "Usunięto sędziego ze stanowiska")
-      }, err => {console.log(err)}).finally(() => {this.loadingReferee = false});
-    }
-  }
-
-  onAddReferee() {
-    if(this.isFormRefereeValid) {
-      let stanowisko_id = this.formReferee.get("position")?.value;
-      this.loadingReferee = true;
-      this.positionsService.addRefereeToPosition(stanowisko_id,this.user.uzytkownik_uuid).then(succes => {
-        this.ui.showFeedback("succes", "Dodano sędziego")
-      }, err => {console.log(err)}).finally(() => {this.loadingReferee = false; this.formReferee.reset()});
-    }
-  }
-
-  stringToArray(string: string): Array<any> {
-    return string ? string.split(', ') : []
-  }
-
   get isFormPostalCodeValid() {
     return this.formPostal.valid
   }
@@ -160,10 +110,6 @@ export class CompetitorComponent {
     return this.formUserType.valid
   }
 
-  get isFormRefereeValid () {
-    return this.formReferee.valid;
-  }
-
   get isLoading() {
     return this.loading;
   }
@@ -171,12 +117,6 @@ export class CompetitorComponent {
   get isLoadingStarterpack() {
     return this.loadingStarerpack;
   }
-
-  get isLoadingReferee () {
-    return this.loadingReferee;
-  }
-
-
 
   openRobotDetails(robot_uuid: any) {
     if(this.userSerceice.isReferee) window.open(`/competitor-zone/(outlet:robot/${robot_uuid})`);
@@ -208,20 +148,6 @@ export class CompetitorComponent {
     let tshirtSizes = this.authService.tshirtSizes ? Object.assign(this.authService.tshirtSizes): undefined;
     return tshirtSizes && this.user ? (tshirtSizes as Array<string>)[this.user?.rozmiar_koszulki-1] : undefined;
   }
-
-  public get positionsOptions(): string | undefined {
-    if (this.allPositions) {
-      // let postions = this.
-      let filtered = this.allPositions.filter(pos => this.allReferePositions.findIndex(el => el == pos) < 0)
-      return JSON.stringify(filtered.map((position: any) => {
-        return { value: position.nazwa_stanowiska, id: position.stanowisko_id };
-      }));
-    }
-    else {
-      return undefined;
-    }
-  }
-
 
   get userTypes(): string
   {
