--- src/app/competitor-zone/referee/positions/positions.component.html
@@ -1,49 +1,0 @@
-<div class="router-child">
-    <section *ngFor="let position of allPositions" class="position">
-        <h1 *ngIf="position.stanowisko_id != editingPosition">{{position.nazwa_stanowiska}} <i class="las la-edit" (click)="editPosition(position.stanowisko_id)"></i></h1>
-        <div class="name" *ngIf="position.stanowisko_id == editingPosition">
-            <app-input [group]="formName" controlName="position"></app-input>
-            <button [disabled]="!isFormNameValid || isLoadingName" class="submit-button" (click)="onUpdateName()"><i class="las la-save"></i></button>
-            <button [disabled]="isLoadingName" class="submit-button" (click)="onDeletePosition()"><i class="las la-trash-alt"></i></button>
-        </div>
-        <p>Kategorie</p>
-        <div class="grid-categories-new" *ngIf="position.stanowisko_id == editingPosition">
-            <app-select *ngIf="categoriesOptions !== undefined && !isLoadingCategories" [group]="formCategory" [collapsed]="true" controlName="category" translatePrefix="competitor-zone.robot" [options]="categoriesOptions"></app-select>
-            <app-spinner *ngIf="categoriesOptions === undefined || isLoadingCategories"></app-spinner>
-            <button [disabled]="!isFormCategoriesValid || isLoadingCategories" class="submit-button grid-button" (click)="onAddCategory()"><i class="las la-plus-circle"></i></button>
-        </div>
-        <div class="grid-categories-old-parent">
-            <div *ngFor="let cat of stringToArray(position.kategorie)" class="grid-categories-old">
-                <div>
-                    <i class="las la-chevron-right"></i>
-                    <label>{{getCategoryName(cat)}}</label>
-                </div>
-                <button class="submit-button grid-button" *ngIf="position.stanowisko_id == editingPosition" (click)="onRemoveCategory(cat)"><i class="las la-trash-alt" ></i></button>
-
-            </div>
-        </div>
-        <p>Sędziowie</p>
-        <div class="grid-categories-new" *ngIf="position.stanowisko_id == editingPosition">
-            <app-select *ngIf="refereeOptions !== undefined && !isLoadingReferee" [group]="formReferee" [collapsed]="true" controlName="referee" translatePrefix="controls" [options]="refereeOptions"></app-select>
-            <app-spinner *ngIf="refereeOptions === undefined || isLoadingReferee"></app-spinner>
-            <button [disabled]="!isFormRefereeValid || isLoadingReferee" class="submit-button grid-button" (click)="onAddReferee()"><i class="las la-plus-circle"></i></button>
-        </div>
-        <div class="grid-categories-old-parent">
-            <div *ngFor="let user of stringToArray(position.sedziowie)" class="grid-categories-old">
-                <div>
-                    <i class="las la-chevron-right"></i>
-                    <label>{{getUserName(user)}}</label>
-                </div>
-                <button class="submit-button grid-button" *ngIf="position.stanowisko_id == editingPosition" (click)="onRemoveReferee(user)"><i class="las la-trash-alt" ></i></button>
-
-            </div>
-        </div>
-    </section>
-    <ng-container *ngIf="{edit: false} as variable">
-        <button class="submit-button" (click)="editPosition(null); variable.edit = true;" *ngIf="!variable.edit">DODAJ STANOWISKO</button>
-        <div class="new-position" *ngIf="variable.edit">
-            <app-input [group]="formName" controlName="position"></app-input>
-            <button [disabled]="!isFormNameValid || isLoadingName" class="submit-button" (click)="onAddPosition(); variable.edit = false;"><i class="las la-save"></i></button>
-        </div>
-    </ng-container>
-</div>
