+++ src/app/competitor-zone/user/my-robots/robot/robot.component.ts
@@ -509,8 +509,7 @@ export class RobotComponent {
   }
 
   get getCategoryFigths() {
-    const fights =  this.robotFights?.filter(el => el.kategoria_id === this.selectedCategory).sort((a, b) => new Date(a.czas_zakonczenia).getTime() - new Date(b.czas_zakonczenia).getTime());
-    return fights
+    return this.robotFights?.filter(el => el.kategoria_id === this.selectedCategory).sort((a, b) => b.walka_id - a.walka_id).sort((a, b) => a.czas_zakonczenia - b.czas_zakonczenia);
   }
 
   get getGroupFigths() {
