+++ src/app/competitor-zone/user/settings/settings.component.html
@@ -52,19 +52,7 @@
             <label><br></label>
         </section>
         <section *ngIf="userService.isReferee">
-            <a>{{'competitor-zone.settings.referee-settings'  | translate}}</a>
-            <label class="checkbox-container">
-              <a [ngClass]="{'checkbox-checked': isDisplayDevice}">{{'competitor-zone.settings.is-sliding' | translate}}</a>
-              <input type="checkbox" (click)="onChangeDisplayDevice()" [checked]="isDisplayDevice">
-              <span class="checkmark"></span>
-          </label>
-            <label class="checkbox-container">
-              <a [ngClass]="{'checkbox-checked': esp32Service.isRFIDDevice}">{{'competitor-zone.settings.is-esp32' | translate}}</a>
-              <input type="checkbox" (click)="esp32Service.onChangeRFIDDevice()" [checked]="esp32Service.isRFIDDevice">
-              <span class="checkmark"></span>
-          </label>
-        </section>
-        <!-- <section *ngIf="userService.isReferee">
+            <a>asdasdjashd</a>
             <a>{{'competitor-zone.settings.test'  | translate}}</a>
             <a *ngIf="returnedPayload">payload: {{returnedPayload}}</a>
             <div class="row">
@@ -79,7 +67,7 @@
                 <button class="submit-button" [disabled]="!isFormGroupRFIDValid" (click)="onWriteRFIDTag()"><i class="las la-save"></i></button>
             </div>
             <label><br></label>
-        </section> -->
+        </section>
     </div>
 
 </div>
