+++ src/assets/i18n/en.json
@@ -188,12 +188,11 @@
                 { "id": 11, "patreonCategory": 5, "description": "Mayor of the City of Rzeszów", "image": "../../assets/jpg/rzeszow.png", "linkToSite": "https://bip.erzeszow.pl/" },
                 { "id": 12, "patreonCategory": 6, "description": "zrobotyzowany.pl", "image": "../../assets/jpg/zrobotyzowany.png", "linkToSite": "https://zrobotyzowany.pl/" },
                 { "id": 13, "patreonCategory": 2, "description": "MTU", "image": "../../assets/jpg/mtu.png", "linkToSite": "https://www.mtu.de/about-us/mtu-worldwide/mtu-aero-engines-polska/website-polska/" },
-                { "id": 14, "patreonCategory": 3, "description": "FANUC", "image": "../../assets/jpg/fanuc.png", "linkToSite": "https://www.fanuc.eu/uk/en" }
-                { "id": 15, "patreonCategory": 3, "description": "ONT", "image": "../../assets/jpg/ont.png", "linkToSite": "https://www.ont.com.pl/" },
-                { "id": 16, "patreonCategory": 3, "description": "Matlab", "image": "../../assets/jpg/mathworks.png", "linkToSite": "https://www.mathworks.com/" },
-                { "id": 17, "patreonCategory": 3, "description": "Zespół Szkół Technicznych", "image": "../../assets/jpg/zst.png", "linkToSite": "http://www.zstrzeszow.pl/" },
-                { "id": 18, "patreonCategory": 6, "description": "Botland", "image": "../../assets/jpg/botland.webp", "linkToSite": "https://botland.com.pl/" },
-                { "id": 19, "patreonCategory": 5, "description": "Voivode of Podkarpacie", "image": "../../assets/jpg/logo_woj_pod.jpg", "linkToSite": "https://rzeszow.uw.gov.pl/wojewoda/wojewoda/" }
+                { "id": 14, "patreonCategory": 3, "description": "ONT", "image": "../../assets/jpg/ont.png", "linkToSite": "https://www.ont.com.pl/" },
+                { "id": 15, "patreonCategory": 3, "description": "Matlab", "image": "../../assets/jpg/mathworks.png", "linkToSite": "https://www.mathworks.com/" },
+                { "id": 16, "patreonCategory": 3, "description": "Zespół Szkół Technicznych", "image": "../../assets/jpg/zst.png", "linkToSite": "http://www.zstrzeszow.pl/" },
+                { "id": 17, "patreonCategory": 6, "description": "Botland", "image": "../../assets/jpg/botland.webp", "linkToSite": "https://botland.com.pl/" },
+                { "id": 18, "patreonCategory": 5, "description": "Voivode of Podkarpacie", "image": "../../assets/jpg/logo_woj_pod.jpg", "linkToSite": "https://rzeszow.uw.gov.pl/wojewoda/wojewoda/" }
 
 
             ],
@@ -325,10 +324,7 @@
             "statistics": "STATISTICS",
             "register": "REGISTER",
             "counting": "VISITORS",
-            "fights": "FIGHT PLANNER",
-            "positions": "POSITIONS",
-            "announcements": "ANNOUNCEMENTS",
-
+            "fights": "WALK PLANNER",
             "hello": {
                 "1": "Welcome to the ",
                 "2": "Competitor Zone!",
@@ -419,9 +415,6 @@
             "get-rfid": "Read RFID",
             "get-lap-time": "Read lap time",
             "set-rfid": "Save data to RFID",
-            "referee-settings": "Device settings",
-            "is-sliding": "Auto sliding results view",
-            "is-esp32": "Display the RFID and gateway control panel (computer only)",
             "change": "* Preferences cannot be changed independently after registration. Please contact the administrator to make changes.",
             "errors": {
                 "succesfuly-added-phone": "Phone number changed successfully",
@@ -446,8 +439,6 @@
                 "3": "Call the robot constructors"
             },
             "next": "Next",
-            "skip": "Skip",
-
             "2": {
                 "1": "Now you can start the fight. When finished, enter the result.",
                 "2": "Enter the number of rounds won by"
@@ -508,8 +499,7 @@
                 "name": "First and last name",
                 "user-uuid": "User UUID",
                 "positions": "Positions",
-                "robots-uuid": "Robot UUID",
-                "robots-name": "Robot name",
+                "robots-uuid": "Robots UUID",
                 "phone": "Phone number",
                 "email": "Email",
                 "categories": "Categories"
@@ -559,10 +549,6 @@
         },
         "categories": {
             "join": "Join now"
-        },
-        "fights-view": {
-            "3rd-fight": "Fight for 3rd place",
-            "no-groups": "No groups"
         }
     },
     "app-rules": {
@@ -1110,25 +1096,6 @@
             "validators": {
                 "required": "This field is required"
             }
-        },
-        "next_fight": {
-            "name": "Next fight",
-            "validators": {
-                "required": "This field is required"
-            }
-        },
-        "referee": {
-            "name": "Referee",
-            "validators": {
-                "required": "This field is required"
-            }
-        },
-        "group_name": {
-            "name": "Group name",
-            "validators": {
-                "required": "This field is required"
-            }
-
         }
     },
     "errors": {
